Changed stylings on stock pages and dashboard.
I decided to use a full width histogram and price to buy and sell since they are more important for the user.
General formatting changes so reading the wall of text should be easier on the eyes.

Changed the order around on the dashboard. The wishlist being near the top made it confusing, sometimes I forget
it's the wishlist and not stocks I have purchased.

PIE graph now has a legend that links to currently owned stock pages for easy navigation. Watchlist should also
link to appropriate stock pages.